requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.60.0
h5py>=3.1.0
Pillow>=8.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
pycocotools>=2.0.4
wandb>=0.12.0
opencv-python>=4.5.0 
captum==0.8.0
openclip